requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2
websockets==13.1

faster-whisper==1.0.3
pyannote.audio==3.3.2

numpy==1.26.4
soundfile==0.12.1
ffmpeg-python==0.2.0
webrtcvad==2.0.10
